api/requirements.txt
Flask==1.1.1
flask_restful==0.3.7
gunicorn==19.9.0
flask_sqlalchemy==2.4.1
pyjwt==1.7.1
Flask-Mail==0.9.1
flask-socketio==4.2.1
eventlet==0.25.1
Flask-Caching==1.7.2
pycrypto==2.6.1
simple-crypt==4.1.7
apscheduler==3.6.1
python-dateutil==2.8.0
psycopg2==2.8.3
boto3==1.9.235
Pillow==6.1.0
pytest==5.0.1
supervisor==4.0.4
